Extract DataOutput scss
Also renamed this to EarningsBreakdown. Both this and the ISAOutput component (which should be renamed as well) need to be refactored. Specifically, there should exist a higher-level component whose sole responsibility is to switch between "cards" of content, i.e. a Tab UI pattern.
